adds module api hooks for resetting and stopping
The difference between reset and stop is important.
You can reset an application at any time, which 
only means that the data is reset like with
meter reset. Stopping an app is more drastic and
also stops any reactive observers etc. (which might
not be what you want for testing).