Merges supplied Module Configuration with current, passing this to the Required Module on initialization
This allows access to the Configuration property of a parent module
during the ‘initialize’ hooks, which is required to wire up the
module’s features without needing to tear down in the stop event. With
this in place we can now abstract a lot of setup, and provide a
configuration API for reliable module implementations.